+++ libs/ionic-react/src/migrations/update-2-0-0/update-2-0-0.ts
@@ -12,15 +12,17 @@ function displayInformation(host: Tree, context: SchematicContext) {
   const config = readJsonInTree(host, 'package.json');
   if (config.devDependencies && config.devDependencies['@nrwl/cypress']) {
     context.logger.info(stripIndents`
-    @testing-library/cypress is being updated to 6.0.0 and includes breaking changes.
-    
-    https://github.com/testing-library/cypress-testing-library/releases/tag/v6.0.0
+      @testing-library/cypress is being updated to 6.0.0 and includes breaking changes.
+      
+      https://github.com/testing-library/cypress-testing-library/releases/tag/v6.0.0
 
-    @testing-library/user-event is being updated to 10.0.1 and includes breaking changes.
+      ---------------------------------------------------------------------------------------
 
-    https://github.com/testing-library/user-event/releases/tag/v9.0.0
-    https://github.com/testing-library/user-event/releases/tag/v10.0.0
-  `);
+      @testing-library/user-event is being updated to 10.0.1 and includes breaking changes.
+
+      https://github.com/testing-library/user-event/releases/tag/v9.0.0
+      https://github.com/testing-library/user-event/releases/tag/v10.0.0
+    `);
   }
 }