+++ libs/ionic-react/CHANGELOG.md
@@ -11,6 +11,7 @@
 - upgrade `@testing-library/user-event` to 10.0.1
 - honor `unitTestRunner` flag
 - set `@nxtend/ionic-react` as the default collection if one is not set when generating an application
+- honor `skipFormat` flag
 
 # 1.0.2